+++ render_lib/include/engine/eng_PerspectiveCamera.h
@@ -36,7 +36,7 @@ public:
 
     virtual ~PerspectiveCamera() = default;
 
-    [[nodiscard]] virtual Ray GenerateRay(int i, int j) override;
+    [[nodiscard]] virtual Ray GenerateRay(float i, float j) override;
 private:
     Vector3DF m_position;
     Vector3DF m_U, m_V, m_W;
